reorganizing modules org/locations
