+++ src/alpss/carrier/filter.py
@@ -1,7 +1,7 @@
 import numpy as np
 from scipy.fft import fft, ifft, fftfreq
 from scipy.fftpack import fftshift
-from alpss.utils.helpers import stft
+from alpss.utils.stft import stft
 from scipy.optimize import curve_fit
