+++ src/alpss/plotting/plots.py
@@ -2,7 +2,7 @@ import matplotlib.pyplot as plt
 from matplotlib.patches import Rectangle
 import pandas as pd
 import os
-from alpss.utils.helpers import stft
+from alpss.utils.stft import stft
 import numpy as np
 import random
 import string
